API_Marketplace/requirements.txt
torch==2.9.1
transformers==4.57.3
sentence-transformers==5.1.2
huggingface-hub==0.35.0
bitsandbytes==0.48.2
accelerate==1.12.0
faiss-cpu==1.13.1
langchain-community==0.4.0
pandas==2.3.3
datamart-profiler==0.12
fastapi==0.124.0
uvicorn==0.38.0
python-multipart==0.0.20
python-dotenv==1.0.1
pycryptodome>=3.20.0
web3>=6.15.1
eth-account>=0.11.0
requests>=2.31.0